==> Project/03_neuralnetwork.ipynb <==
# Neural network

import os
import keras
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from keras.models import Sequential
from keras.layers import Dense, Dropout
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_log_error
import matplotlib


X_train = pd.read_csv("./dataset/X_train.csv", header=None)
X_test = pd.read_csv("./dataset/X_test.csv", header=None)
y_train = pd.read_csv("./dataset/y_train.csv", header=None)

y_train_log = np.log(y_train)

inp_size = X_train.shape[1]

model = Sequential()
model.add(Dense(512, input_dim=inp_size, kernel_initializer='normal',
                activation='relu'))
model.add(Dropout(0.3))
model.add(Dense(256, kernel_initializer='normal', activation='relu'))
model.add(Dense(128, kernel_initializer='normal', activation='relu'))
model.add(Dropout(0.3))

model.add(Dense(1, kernel_initializer='normal', activation='relu'))

model.compile(loss='mean_squared_error', 
              optimizer=keras.optimizers.Nadam())
model.fit(X_train, y_train_log, epochs=2000, validation_split=0.1)

y_pred = model.predict(X_train)
fig, ax = plt.subplots(figsize=(15,12))
plt.style.use('ggplot')
plt.plot(y_pred, y_train_log, 'ro')
plt.xlabel('Prediction')
plt.ylabel('Reality')
ax.plot([y_train_log.min(), y_train_log.max()], 
        [y_train_log.min(), y_train_log.max()], 'k--')
plt.show()

y_pred = np.exp(model.predict(X_test)[:,0])
result=pd.DataFrame({'Id':range(1461, 2920), 'SalePrice':y_pred})
result.to_csv("dataset/submission2.csv",index=False)
